+++ src/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import './reset.css';
 import './App.css';
-import NameModule from './components/NameModule';
+import NameModule from './components/NameModule/NameModule';
 // import TestModule from './components/TestModule';
 
 class App extends Component {
