+++ src/components/NameModule/NameModule.scss
@@ -21,6 +21,7 @@ $outer-border: 0px solid #9e9e9e;
     "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
     sans-serif;
     margin-bottom: 2px;
+    cursor: pointer;
   }
   .translated {
     height: 16px;
