+++ src/components/NameModule/NamePoolModal.scss
