+++ src/components/NameModule/partials/trinarySliderList.js
